+++ app/upgrades.go
@@ -3,6 +3,15 @@ package app
 import (
 	"fmt"
 
+	_ "github.com/cosmos/cosmos-sdk/client/docs/statik"
+	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
+	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
+	crisistypes "github.com/cosmos/cosmos-sdk/x/crisis/types"
+	distrtypes "github.com/cosmos/cosmos-sdk/x/distribution/types"
+	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
+	govv1types "github.com/cosmos/cosmos-sdk/x/gov/types/v1"
+	slashingtypes "github.com/cosmos/cosmos-sdk/x/slashing/types"
+	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 	upgradetypes "github.com/cosmos/cosmos-sdk/x/upgrade/types"
 
 	"github.com/NibiruChain/nibiru/v2/app/upgrades"
@@ -32,13 +41,33 @@ var Upgrades = []upgrades.Upgrade{
 }
 
 func (app *NibiruApp) setupUpgrades() {
+	// see https://github.com/cosmos/cosmos-sdk/blob/666c345ad23ddda9523cc5cd1b71187d91c26f34/simapp/upgrades.go#L35-L57
+	for _, subspace := range app.paramsKeeper.GetSubspaces() {
+		switch subspace.Name() {
+		case authtypes.ModuleName:
+			subspace.WithKeyTable(authtypes.ParamKeyTable()) //nolint:staticcheck
+		case banktypes.ModuleName:
+			subspace.WithKeyTable(banktypes.ParamKeyTable()) //nolint:staticcheck
+		case stakingtypes.ModuleName:
+			subspace.WithKeyTable(stakingtypes.ParamKeyTable()) //nolint:staticcheck
+		case distrtypes.ModuleName:
+			subspace.WithKeyTable(distrtypes.ParamKeyTable()) //nolint:staticcheck
+		case slashingtypes.ModuleName:
+			subspace.WithKeyTable(slashingtypes.ParamKeyTable()) //nolint:staticcheck
+		case govtypes.ModuleName:
+			subspace.WithKeyTable(govv1types.ParamKeyTable()) //nolint:staticcheck
+		case crisistypes.ModuleName:
+			subspace.WithKeyTable(crisistypes.ParamKeyTable()) //nolint:staticcheck
+		}
+	}
+
 	app.setUpgradeHandlers()
 	app.setUpgradeStoreLoaders()
 }
 
 func (app *NibiruApp) setUpgradeHandlers() {
 	for _, u := range Upgrades {
-		app.upgradeKeeper.SetUpgradeHandler(u.UpgradeName, u.CreateUpgradeHandler(app.ModuleManager, app.configurator, app.ibcKeeper.ClientKeeper))
+		app.upgradeKeeper.SetUpgradeHandler(u.UpgradeName, u.CreateUpgradeHandler(app.ModuleManager, app.Configurator(), app.ibcKeeper.ClientKeeper))
 	}
 }