+++ eth/crypto/codec/amino.go
@@ -4,7 +4,6 @@ package codec
 import (
 	"github.com/cosmos/cosmos-sdk/codec"
 	"github.com/cosmos/cosmos-sdk/codec/legacy"
-	cryptocodec "github.com/cosmos/cosmos-sdk/crypto/codec"
 	"github.com/cosmos/cosmos-sdk/crypto/keyring"
 
 	"github.com/NibiruChain/nibiru/v2/eth/crypto/ethsecp256k1"
@@ -19,7 +18,6 @@ func RegisterCrypto(cdc *codec.LegacyAmino) {
 		ethsecp256k1.PrivKeyName, nil)
 
 	keyring.RegisterLegacyAminoCodec(cdc)
-	cryptocodec.RegisterCrypto(cdc)
 
 	// NOTE: update SDK's amino codec to include the ethsecp256k1 keys.
 	// DO NOT REMOVE unless deprecated on the SDK.