+++ x/sudo/cli/cli_test.go
@@ -24,7 +24,6 @@ import (
 	"github.com/NibiruChain/nibiru/v2/x/common/set"
 	"github.com/NibiruChain/nibiru/v2/x/common/testutil"
 	"github.com/NibiruChain/nibiru/v2/x/common/testutil/genesis"
-	"github.com/NibiruChain/nibiru/v2/x/common/testutil/testapp"
 	"github.com/NibiruChain/nibiru/v2/x/common/testutil/testnetwork"
 	"github.com/NibiruChain/nibiru/v2/x/sudo/cli"
 )
@@ -107,9 +106,8 @@ func TestSuite_IntegrationSuite_RunAll(t *testing.T) {
 
 func (s *TestSuite) SetupSuite() {
 	testutil.BeforeIntegrationSuite(s.T())
-	testapp.EnsureNibiruPrefix()
 
-	genState := genesis.NewTestGenesisState(app.MakeEncodingConfig())
+	genState := genesis.NewTestGenesisState(app.MakeEncodingConfig().Codec)
 	genState, rootPrivKey, rootAddr := genesis.AddSudoGenesis(genState)
 	s.root = Account{
 		privKey:    rootPrivKey,