+++ x/sudo/keeper/msg_server_test.go
@@ -2,7 +2,6 @@ package keeper_test
 
 import (
 	"testing"
-	"time"
 
 	"github.com/NibiruChain/nibiru/v2/x/sudo/keeper"
 
@@ -19,12 +18,8 @@ import (
 	"github.com/NibiruChain/nibiru/v2/x/sudo"
 )
 
-func init() {
-	testapp.EnsureNibiruPrefix()
-}
-
 func setup() (*app.NibiruApp, sdk.Context) {
-	return testapp.NewNibiruTestAppAndContextAtTime(time.Now().UTC())
+	return testapp.NewNibiruTestAppAndContext()
 }
 
 func TestGenesis(t *testing.T) {